# [blas] Optimize gemv and gemm for float16 data type