@skykongkong8 2024-02-05
> > Great work! Suggestion : typical evaluation of BLAS library is conducted with [FLOPS](https://cs.stanford.edu/people/shadjis/blas.html) how about adding this? Question: is ms in your test result refers to milliseconds? (or microseconds?)
> 
> for sure. there's still room for optimization if it's compared to sgemv and sgemm. I'll also check with FLOPS in later optimization. also, it is milliseconds. will mark as μs when it comes to micro :)

Thanks a lot! 👍